execution/script_exemplo.py
#!/usr/bin/env python3
"""
Script de exemplo para demonstrar a estrutura de execução.

Este script é determinístico e confiável. Ele lida com operações específicas
sem depender de lógica probabilística de LLMs.
"""

import os
from dotenv import load_dotenv

# Carrega variáveis de ambiente
load_dotenv()


def exemplo_funcao():
    """
    Função de exemplo que demonstra como estruturar scripts de execução.
    
    Returns:
        str: Mensagem de sucesso
    """
    # Lógica determinística aqui
    resultado = "Processamento concluído"
    return resultado


if __name__ == "__main__":
    print("Executando script de exemplo...")
    resultado = exemplo_funcao()
    print(f"Resultado: {resultado}")
